+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [40.1.1] 2026-02-19
+### Changed
+- Switched to ESLint, updated npm packages, and adjusted Cypress tests.
+### Fixed
+- Prevent adding duplicate unit options when a quantity initial value
+  unit matches an existing unitOption.
+- Fixed a cypress test that depended on unreliable data.
+
+## [40.1.0] 2026-01-22
+### Fixed
+- A Questionnaire with multiple initial values on a repeating question
+  should render multiple questions, each with a single default answer.
+- The same form, when exported back to Questionnaire, should contain
+  only one question with an array of initial values.
+
 ## [40.0.0] 2026-01-22
 ### Added
 - Support for observationExtractEntry extension in Observation extract.